requirements.txt
asgiref
Django
django-cors-headers
django-debug-toolbar
django-filter
djangorestframework
pytz
sqlparse
psycopg2-binary
python-dotenv
Pillow
reportlab
openpyxl
validate-docbr
python-decouple
pandas
pytesseract==0.3.10
PyPDF2==3.0.1
requests==2.31.0
pdf2image==1.17.0